create components/i18n.ts to prevent SSR build errors
